Refactor: 팀 카드 공용 컴포넌트로 사용
